refactor(enterprise/coderd): split AI provider keys handlers into ai_provider_keys.go
The keys handlers are a distinct sub-resource; keep them in their
own file to make the providers file easier to navigate.
